real time data
in this commit I was able to get my server to send real time data to the client (for the RPi). I was able to do this by removing the c.close() which would close after sending data once (I also added time.sleep to slow down the data being sent). The next thing I had to do was add a loop in the client code to keep checking for data.