added real time updates
In this commit I have added the ability to see real time updates. I have done this by adding a loop to keep checking for the data recv and then decoding it and finally printing it (with a small delay on the server side). I have also added the ability to exit the program using ctrl-c. If the client fails to connect for whatever reason the code will also now print "lost connection".