fixed re-connection issue
In this commit I have added a try except to detect the error when the client disconnects. If the error occurs then it will print a message saying that the client has disconnected and will c.close() to stop sending data until someone connects again. I also added a keyboard interrupt.